feat: add support for cert auth
Add two new flags, -cert-file and -key-file. They can be used to
configure to the broker using mutual TLS authentication. The flags are
mutually exclusive with each other. Only one of "username and password"
or "cert-file and key-file" may be used to configure authentication. If
neither are given, then an anonymous connection is attempted.